ov006/dScMgHanachan_c: extend the partial from 22 of 61 to 49 of 61
`#pragma defer_codegen off` dissolves the pragma contradiction that cut this
TU at 22 members, and the whole larger side of the sourceless hole now
reproduces in one compile.

The 22-function version stopped where it did because four members of the
49-function side reproduce only with `#pragma opt_strength_reduction off` and
three only with it on, while Render wants `opt_common_subs off`, which costs
five others.  Those pragmas were measured to be file-global last-wins with no
positional escape.  That measurement holds only while codegen is deferred:
with `#pragma defer_codegen off` mwccarm 2004/b56 generates each function as it
is parsed and both pragmas bind positionally, so the four strength-reduction
members carry their own push/pop bracket, Render carries its own, and all 49
match at once.

The same pragma flips emission order from reverse-source to source order, so
the file is now written lowest ROM address first.  The two are one edit --
flipping either alone fails `linkcheck [4b/8]`'s ROM-ascending refusal.  Both
directions were measured on the identical 22-member set before the fold-back.

Three merge-only byte diffs, none pragma-related, were fixed at the source:
three whole-struct copies of a two-int pair needed the array-spelled shadow
type because C++ scalarizes what the legacy C block-moved, and
OnYoshiTryEat's read-modify-write of +0xbc needed a `const` on the read to stop
mwcc CSEing the field address into its own register.

  tubuild verify : 49/49 MATCH, objisolate clean, reloc-destinations clean,
                   emission order ROM-ascending -> TEXT-VERIFIED

Only the smaller side of the sourceless hole (ordinals 0..10, including the
destructor pair) stays out; that edge is unchanged and is not a destructor-
order refusal.
