+++ config/arm9/overlays/ov006/delinks.txt
@@ -2835,116 +2835,9 @@ src/func_ov006_020ea8e0.c:
     complete
     .text start:0x020ea8e0 end:0x020ea914
 
-src/func_ov006_020eac38.c:
-    complete
-    .text start:0x020eac38 end:0x020eb018
-
 src/actors/dScMgHanachan_c.cpp:
     complete
-    .text start:0x020eb018 end:0x020ec458
-
-src/func_ov006_020ec458.c:
-    complete
-    .text start:0x020ec458 end:0x020ec4dc
-
-src/func_ov006_020ec4dc.c:
-    complete
-    .text start:0x020ec4dc end:0x020ec6e8
-
-src/func_ov006_020ec6e8.c:
-    complete
-    .text start:0x020ec6e8 end:0x020ec84c
-
-src/func_ov006_020ec84c.c:
-    complete
-    .text start:0x020ec84c end:0x020ec93c
-
-src/func_ov006_020ec93c.c:
-    complete
-    .text start:0x020ec93c end:0x020ec9c0
-
-src/func_ov006_020ec9c0.c:
-    complete
-    .text start:0x020ec9c0 end:0x020ecb80
-
-src/func_ov006_020ecb80.c:
-    complete
-    .text start:0x020ecb80 end:0x020ecba4
-
-src/func_ov006_020ecba4.c:
-    complete
-    .text start:0x020ecba4 end:0x020ecdb8
-
-src/func_ov006_020ecdb8.c:
-    complete
-    .text start:0x020ecdb8 end:0x020ecec8
-
-src/_ZN15dScMgHanachan_c16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x020ecec8 end:0x020ecee4
-
-src/_ZN15dScMgHanachan_c6RenderEv.cpp:
-    complete
-    .text start:0x020ecee4 end:0x020ed18c
-
-src/_ZN15dScMgHanachan_c8BehaviorEv.cpp:
-    complete
-    .text start:0x020ed18c end:0x020ed270
-
-src/func_ov006_020ed270.c:
-    complete
-    .text start:0x020ed270 end:0x020ed274
-
-src/func_ov006_020ed274.c:
-    complete
-    .text start:0x020ed274 end:0x020ed300
-
-src/func_ov006_020ed300.c:
-    complete
-    .text start:0x020ed300 end:0x020ed328
-
-src/func_ov006_020ed328.c:
-    complete
-    .text start:0x020ed328 end:0x020ed32c
-
-src/func_ov006_020ed32c.c:
-    complete
-    .text start:0x020ed32c end:0x020ed34c
-
-src/func_ov006_020ed34c.c:
-    complete
-    .text start:0x020ed34c end:0x020ed40c
-
-src/func_ov006_020ed40c.c:
-    complete
-    .text start:0x020ed40c end:0x020ed494
-
-src/func_ov006_020ed494.c:
-    complete
-    .text start:0x020ed494 end:0x020ed81c
-
-src/func_ov006_020ed81c.c:
-    complete
-    .text start:0x020ed81c end:0x020ed844
-
-src/func_ov006_020ed844.cpp:
-    complete
-    .text start:0x020ed844 end:0x020ed8a4
-
-src/func_ov006_020ed8a4.c:
-    complete
-    .text start:0x020ed8a4 end:0x020eda48
-
-src/_ZN15dScMgHanachan_c13OnYoshiTryEatEi.cpp:
-    .text start:0x020eda48 end:0x020edb04
-
-src/_ZN15dScMgHanachan_c13InitResourcesEv.cpp:
-    complete
-    .text start:0x020edb04 end:0x020edcb0
-
-src/func_ov006_020edcb0.c:
-    complete
-    .text start:0x020edcb0 end:0x020ede18
+    .text start:0x020eac38 end:0x020ede18
 
 src/minigames/d_s_mg_hanachan.cpp:
     complete
