--- src/func_ov006_020ecb80.c
@@ -1,6 +1,0 @@
-extern void _Z14ApproachLinearRsss(short *, short, int);
-void func_ov006_020ecb80(void *a)
-{
-  char *p = (char *) a;
-  _Z14ApproachLinearRsss((short *) (p + 0x7a), (short) ((*((short *) (p + 0x76))) - 0x2000), 0x100);
-}
