--- src/func_ov006_020ed328.c
@@ -1,3 +1,0 @@
-void func_ov006_020ed328(void)
-{
-}
